Break configure into its own module
