+++ tests/test_restore.py
@@ -5,7 +5,8 @@ import vault_dev
 import docker
 import privateer2.restore
 from privateer2.config import read_config
-from privateer2.keys import configure, keygen_all
+from privateer2.configure import configure
+from privateer2.keys import keygen_all
 from privateer2.restore import restore
